+++ cmd/configloader/main.go
@@ -35,10 +35,17 @@ func main() {
 				Path: "/etc/configloader.yml",
 				// CODE_REVIEW_CATCH_ME: should probalby make this a factory method for
 				// all types of unmarshaler that takes an opts or something... dunno
-				Unmarshal: config.YamlValueTemplateUnmarshal[map[any]any](nil),
+				Unmarshal: config.
+					YamlValueTemplateUnmarshal[map[any]any](nil),
 			},
 			config.DirSource[map[any]any]{Path: "/etc/configloader.d"},
-			config.FileSource[map[any]any]{Path: "~/.config/configloader.yml"},
+			config.FileSource[map[any]any]{
+				Path: "~/.config/configloader.yml",
+				// CODE_REVIEW_CATCH_ME: should probalby make this a factory method for
+				// all types of unmarshaler that takes an opts or something... dunno
+				Unmarshal: config.
+					YamlValueTemplateUnmarshal[map[any]any](nil),
+			},
 			config.DirSource[map[any]any]{Path: "~/.config/configloader.d"},
 		},
 	}
